change pm25 type to int to reflect official documentation

+++ ambient.go
@@ -62,7 +62,7 @@ type Record struct {
 	Humidityin     int
 	LastRain       time.Time
 	Maxdailygust   float64
-	Pm25           float64
+	Pm25           int
 	Pm25_24h       float64
 	Relay1         int
 	Relay2         int
